Latest update

+++ project_3/src/Store/Staff/Jobs/CheckRegister.java
@@ -5,7 +5,7 @@ import Store.Staff.*;
 
 public class CheckRegister extends Job {
     public void do_job(Store s, Staff p) {
-        double cash = s.getCashReg();
+        float cash = s.getCashReg();
         System.out.println("Register has $"+cash);
         if (cash<75) {
             GoToBank next = new GoToBank();
